Refractor to seperate lexer buffering from rest of parse cursor
